Enable pipefail flag
This is to force early exit if any command in a pipeline
fails instead of allowing the last exit code in the pipeline
from determining overall success/failure.

Not directly related, but this should prove useful in helping
to narrow down the specific cause of invalid patch reports
as noted on the related ticket.

refs whyaskwhy/email-updates#9